requirements.txt
# Core dependencies
opencv-python==4.11.0.86
numpy==2.2.4
matplotlib==3.10.1
pillow==11.1.0
scikit-image==0.25.2
requests==2.32.3

# Required by matplotlib
contourpy==1.3.1
cycler==0.12.1
fonttools==4.57.0
kiwisolver==1.4.8

# Required by scikit-image
imageio==2.37.0
lazy_loader==0.4
networkx==3.4.2
tifffile==2025.3.30
platformdirs==4.3.8
pooch==1.8.2

# Required by requests
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
urllib3==2.4.0

# Common dependencies
packaging==24.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
six==1.17.0
